advanced_rag_project/requirements.txt
streamlit
google-generativeai
scikit-learn
pandas
numpy
